+++ src/appReset.js
@@ -1,4 +1,4 @@
-
+import { clearDlcLocalStorages } from "./utils/dlcStorage";
 // const tcResetButton = document.getElementById('resetTimeCopy')
 // tcResetButton.addEventListener('click',appReset)
 // <button hidden id="resetTimeCopy">Reset</button>
@@ -27,7 +27,7 @@ document.addEventListener("DOMContentLoaded", function () {
 
 function appReset(){
   console.log('🔄 RESET APP')
-    localStorage.clear();
-    sessionStorage.clear();
-    window.location.reload();
-  }
+  localStorage.clear();
+  sessionStorage.clear();
+  window.location.reload();
+}