+++ src/dlc/filters/None/None.js
@@ -1,0 +1,11 @@
+export function filter_None(clipboardString) {
+
+    let bookingData = []
+    let itemObject = createFilterObject('',
+      '','',clipboardString,
+      '','','',''
+    )
+    bookingData.push(itemObject)
+
+    return bookingData
+}